fix: replace styles, add constants
