+++ src/constants/common.dictionary.ts
@@ -6,3 +6,5 @@ export const PATH = {
 export const MAIN_PAGE_TITLE = 'Main Page Title';
 export const WELCOME_PAGE_TITLE = 'Welcome to Lemasello';
 export const PAGE_404_TITLE = 'Error 404, this page can not be found...';
+export const PAGE_404_MESSAGE = 'Something wrong, you can back to ';
+export const PAGE_404_LINK_TO = 'main page';
